Pagamento/requirements.txt
sqlalchemy
pymysql
kafka-python
pydantic
python-dotenv
